requirements.txt
Django==1.8.6
django-allauth==0.24.1
django-extensions==1.5.9
django-rest-auth==0.5.0
djangorestframework==3.3.1
oauthlib==1.0.3
python-openid==2.2.5
requests==2.8.1
requests-oauthlib==0.5.0
six==1.10.0